requirements.txt
tabulate~=0.8.10
pyodbc~=5.0.1
requests~=2.31.0
simplejson~=3.19.2
apache-airflow~=2.9.1